refactor & enhance UI components: replaced old components with new, reusable ones
