+++ SwimMate/iOSViews/SwimViews/SetView/SetPage.swift
@@ -206,152 +206,6 @@ struct RecommendedSetCard: View {
     }
 }
 
-// MARK: - Modern Set Card
-struct ModernSetCard: View {
-    let swimSet: SwimSet
-    let isFavorite: Bool
-    let toggleFavorite: () -> Void
-    
-    private var difficultyColor: Color {
-        switch swimSet.difficulty {
-        case .beginner: return .green
-        case .intermediate: return .orange
-        case .advanced: return .red
-        }
-    }
-    
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            // Header with difficulty badge
-            HStack {
-                Text(swimSet.difficulty.rawValue.capitalized)
-                    .font(.system(size: 10, weight: .bold))
-                    .foregroundColor(.white)
-                    .padding(.horizontal, 8)
-                    .padding(.vertical, 4)
-                    .background(difficultyColor)
-                    .cornerRadius(6)
-                
-                Spacer()
-                
-                Button(action: toggleFavorite) {
-                    Image(systemName: isFavorite ? "heart.fill" : "heart")
-                        .font(.system(size: 16, weight: .medium))
-                        .foregroundColor(isFavorite ? .red : .gray)
-                }
-                .buttonStyle(PlainButtonStyle())
-            }
-            
-            // Title
-            Text(swimSet.title)
-                .font(.system(size: 16, weight: .bold))
-                .foregroundColor(.primary)
-                .lineLimit(2)
-                .multilineTextAlignment(.leading)
-            
-            // Key metrics
-            VStack(spacing: 6) {
-                HStack {
-                    Image(systemName: "ruler")
-                        .font(.system(size: 12))
-                        .foregroundColor(.secondary)
-                    Text("\(swimSet.totalDistance) \(swimSet.measureUnit.rawValue)")
-                        .font(.system(size: 14, weight: .medium))
-                        .foregroundColor(.secondary)
-                    Spacer()
-                }
-                
-                HStack {
-                    Image(systemName: "list.bullet")
-                        .font(.system(size: 12))
-                        .foregroundColor(.secondary)
-                    Text("\(swimSet.components.count) components")
-                        .font(.system(size: 14, weight: .medium))
-                        .foregroundColor(.secondary)
-                    Spacer()
-                }
-                
-                if let primaryStroke = swimSet.primaryStroke {
-                    HStack {
-                        Image(systemName: "figure.pool.swim")
-                            .font(.system(size: 12))
-                            .foregroundColor(.secondary)
-                        Text(primaryStroke.description)
-                            .font(.system(size: 14, weight: .medium))
-                            .foregroundColor(.secondary)
-                        Spacer()
-                    }
-                }
-            }
-            
-            Spacer()
-        }
-        .padding()
-        .frame(maxWidth: .infinity, minHeight: 160, maxHeight: 180, alignment: .topLeading)
-        .background(Color(UIColor.systemBackground))
-        .cornerRadius(16)
-        .shadow(color: .black.opacity(0.06), radius: 8, x: 0, y: 2)
-        .overlay(
-            RoundedRectangle(cornerRadius: 16)
-                .stroke(difficultyColor.opacity(0.2), lineWidth: 1)
-        )
-    }
-}
-
-// MARK: - Stat Card
-struct StatCard: View {
-    let title: String
-    let value: String
-    let icon: String
-    let color: Color
-    
-    var body: some View {
-        VStack(spacing: 8) {
-            Image(systemName: icon)
-                .font(.system(size: 20, weight: .medium))
-                .foregroundColor(color)
-            
-            Text(value)
-                .font(.system(size: 18, weight: .bold))
-                .foregroundColor(.primary)
-            
-            Text(title)
-                .font(.system(size: 12, weight: .medium))
-                .foregroundColor(.secondary)
-                .multilineTextAlignment(.center)
-        }
-        .frame(maxWidth: .infinity)
-        .padding()
-        .background(Color(UIColor.systemBackground))
-        .cornerRadius(12)
-        .shadow(color: .black.opacity(0.05), radius: 6, x: 0, y: 2)
-    }
-}
-
-// MARK: - Empty State View
-struct EmptyStateView: View {
-    var body: some View {
-        VStack(spacing: 20) {
-            Image(systemName: "magnifyingglass.circle")
-                .font(.system(size: 60, weight: .light))
-                .foregroundColor(.secondary)
-            
-            VStack(spacing: 8) {
-                Text("No Sets Found")
-                    .font(.system(size: 20, weight: .semibold))
-                    .foregroundColor(.primary)
-                
-                Text("Try adjusting your filters or search terms")
-                    .font(.system(size: 16, weight: .regular))
-                    .foregroundColor(.secondary)
-                    .multilineTextAlignment(.center)
-            }
-        }
-        .padding(40)
-        .frame(maxWidth: .infinity)
-    }
-}
-
 // MARK: - Preview
 
 #Preview {
